Move menu_bar_controller.py into RimSort.controllers
